+++ dz19/src/main/resources/templates/books.html
@@ -46,7 +46,7 @@
         <td width="150 " th:text="${book.author.firstName}" >John</td>
         <td width="150" th:text="${book.author.lastName}" >Doe</td>
         <td width="100" th:text="${book.genre.name}" >fantasy</td>
-        <td width="150" ><a th:href="@{/comments/list(id=${book.id})}" href="comments.html">Смотреть/добавить</a></td>
+        <td width="150" ><a th:href="@{/comments(book-id=${book.id})}" href="comments.html">Смотреть/добавить</a></td>
         <td width="50"  ><a th:href="@{/addbook/edit(id=${book.id})}" href="addbook.html">Изменить</a></td>
     </tr>
     </tbody>
