--- dz19/src/test/java/ru/otus/dz19/controller/AddBookControllerTest.java
@@ -1,68 +1,0 @@
-package ru.otus.dz19.controller;
-
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.Mockito;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.web.servlet.MockMvc;
-import ru.otus.dz19.domain.*;
-import ru.otus.dz19.repository.UserRepository;
-import ru.otus.dz19.security.SecurityConfiguration;
-import ru.otus.dz19.service.CommentService;
-import ru.otus.dz19.service.LibraryService;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
-
-import static org.hamcrest.Matchers.containsString;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@RunWith(SpringRunner.class)
-@WebMvcTest(AddBookController.class)
-@WithMockUser(
-        username = "user",
-        authorities = {"ROLE_USER"}
-
-)
-public class AddBookControllerTest {
-
-    @Autowired
-    private MockMvc mvc;
-
-    @MockBean
-    private LibraryService libraryService;
-
-    @MockBean
-    private UserRepository userRepository;
-
-    @MockBean
-    private CommentService commentService;
-
-    @Configuration
-    @ComponentScan(basePackageClasses = {AddBookController.class, SecurityConfiguration.class})
-    public static class TestConf {
-    }
-
-    @Test
-    public void addBookPage() throws Exception {
-        mvc.perform(get("/addbook"))
-                .andExpect(status().isForbidden());
-    }
-
-    @Test
-    public void editBookPage() throws Exception {
-        mvc.perform(get("/addbook/edit?id=1" ))
-                .andExpect(status().isForbidden());
-    }
-}
