+++ dz19/src/test/java/ru/otus/dz19/controller/BookControllerRoleAdminTest.java
@@ -3,6 +3,7 @@ package ru.otus.dz19.controller;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -21,10 +22,14 @@ import ru.otus.dz19.service.LibraryService;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.Optional;
 
+import static org.hamcrest.Matchers.containsString;
 import static org.junit.Assert.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.redirectedUrl;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -67,6 +72,7 @@ public class BookControllerRoleAdminTest {
         genre = new Genre("роман-эпопея");
         genre.setId(1);
         book = new Book("Война и мир", author, genre);
+        book.setId(1);
         comment = new Comment("Эпично, но слишком затянуто.", book);
         books = Arrays.asList(book);
         bookDto = new BookDto(1, "Мертвые души", "Николай", "Гоголь", "поэма");
@@ -98,5 +104,20 @@ public class BookControllerRoleAdminTest {
         mvc.perform(post("/books/add/1")
                 .flashAttr("bookDto", bookDto))
                 .andExpect(redirectedUrl("/books"));
+    }
+
+    @Test
+    public void addBookPage() throws Exception {
+        mvc.perform(get("/addbook"))
+                .andExpect(status().isOk());
+    }
+
+    @Test
+    public void editBookPage() throws Exception {
+        Mockito.when(libraryService.findBookById(1)).thenReturn(Optional.of(book));
+        mvc.perform(get("/addbook/edit?id=" + book.getId()))
+                .andExpect(status().isOk())
+                .andDo(print())
+                .andExpect(content().string(containsString(book.getTitle())));
     }
 }
