+++ dz19/src/test/java/ru/otus/dz19/controller/CommentControllerTest.java
@@ -12,29 +12,26 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
-import ru.otus.dz19.domain.Author;
-import ru.otus.dz19.domain.Book;
-import ru.otus.dz19.domain.Comment;
-import ru.otus.dz19.domain.Genre;
+import ru.otus.dz19.domain.*;
 import ru.otus.dz19.repository.UserRepository;
-import ru.otus.dz19.security.SecurityConfiguration;
 import ru.otus.dz19.service.CommentService;
 import ru.otus.dz19.service.LibraryService;
 
+import java.util.Arrays;
+import java.util.List;
 import java.util.Optional;
 
 import static org.hamcrest.Matchers.containsString;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 @RunWith(SpringRunner.class)
-@WebMvcTest(AddCommentController.class)
+@WebMvcTest(CommentController.class)
 @WithMockUser(
         username = "user"
 )
-public class AddCommentControllerTest {
+public class CommentControllerTest {
 
     @Autowired
     private MockMvc mvc;
@@ -43,14 +40,13 @@ public class AddCommentControllerTest {
     private LibraryService libraryService;
 
     @MockBean
-    private UserRepository userRepository;
-
-    @MockBean
     private CommentService commentService;
 
+    @MockBean
+    private UserRepository userRepository;
 
     @Configuration
-    @ComponentScan(basePackageClasses = {AddCommentController.class})
+    @ComponentScan(basePackageClasses = {CommentController.class})
     public static class TestConf {
     }
 
@@ -58,6 +54,8 @@ public class AddCommentControllerTest {
     private Author author;
     private Genre genre;
     private Book book;
+    private List<Comment> comments;
+    private CommentDto commentDto;
 
     @Before
     public void setUp() throws Exception {
@@ -69,6 +67,37 @@ public class AddCommentControllerTest {
         book.setId(1);
         comment = new Comment("Эпично, но слишком затянуто.", book);
         comment.setId(1);
+        comments = Arrays.asList(comment);
+        commentDto = new CommentDto(1, "Очень понравилось","Война и мир",1);
+    }
+
+    @Test
+    public void commentsByBookPage() throws Exception {
+        Mockito.when(libraryService.findBookById(1)).thenReturn(Optional.of(book));
+        Mockito.when(commentService.listCommentsByBook(book.getId())).thenReturn(comments);
+        mvc.perform(get("/comments?book-id=" + book.getId()))
+                .andExpect(status().isOk());
+    }
+
+    @Test
+    public void saveComment() throws Exception {
+        mvc.perform(post("/comments/add")
+                .flashAttr("commentDto", commentDto))
+                .andExpect(redirectedUrl("/comments?book-id=1"));
+    }
+
+    @Test
+    public void updateComment() throws Exception {
+        mvc.perform(post("/comments/add")
+                .flashAttr("commentDto", commentDto))
+                .andExpect(redirectedUrl("/comments?book-id=1"));
+    }
+
+    @Test
+    public void deleteComment() throws Exception {
+        mvc.perform(post("/comments/add")
+                .flashAttr("commentDto", commentDto))
+                .andExpect(redirectedUrl("/comments?book-id=1"));
     }
 
     @Test
